Implement weather API.

+++ app/src/main/java/com/coldani3/dangoniser/data/api/Coord.kt
@@ -1,6 +1,6 @@
 package com.coldani3.dangoniser.data.api
 
 data class Coord(
-    val lat: Int,
-    val lon: Int
+    val lat: Double,
+    val lon: Double
 )
